+++ src/layouts/Admin.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect } from "react";
 
 import { Outlet } from "react-router-dom";
 
@@ -8,12 +8,22 @@ import Sidebar from "./Sidebar/Sidebar";
 import Header from "./Header/Header";
 
 export default function Admin({ mode }) {
+  const [contentHeight, setContentHeight] = React.useState(window.innerHeight);
+
+  useEffect(() => {
+    function handleResize() {
+      setContentHeight(window.innerHeight);
+    }
+
+    window.addEventListener('resize', handleResize);
+  });
+
   return (
     <>
       <Sidebar mode={mode} />
       <div className="p-4 admin-wrapper">
         <Header />
-        <div className="content-wrapper">
+        <div className="content-wrapper" style={{height: (contentHeight - 114) + 'px'}}>
           <Outlet />
         </div>
       </div>
