+++ src/layouts/Header/Header.js
@@ -1,19 +1,46 @@
-import React from "react";
-import { useLocation } from "react-router-dom";
+import React, { useEffect, useState } from "react";
+import { useLocation, useNavigate } from "react-router-dom";
 import Search from "./Search";
 
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
 export default function Header() {
-  let location = useLocation();
+  const location = useLocation();
+  const navigate = useNavigate();
+
+  const curFinacePage =
+    location.pathname.indexOf("/erp/finance") >= 0
+      ? location.pathname.replace("/erp/finance", "")
+      : "";
+  const [financePage, setFinancePage] = useState(
+    curFinacePage === ""
+      ? "sales"
+      : curFinacePage.replace("/", "").toLowerCase()
+  );
+
+  useEffect(() => {
+    if (
+      location.pathname.indexOf("/erp/finance") >= 0 &&
+      location.pathname !== "/erp/finance/".financePage
+    ) {
+      // navigate("/erp/finance/".financePage);
+    }
+  });
 
   return (
     <div className="fixed admin-header">
-      {location.pathname === "/erp/finance" ? (
-        <select name="finance" className="float-right sel-finance">
-          <option>Sales</option>
-          <option>Expenses</option>
-          <option>Profits</option>
+      {location.pathname.indexOf("/erp/finance") >= 0 ? (
+        <select
+          name="finance"
+          className="float-right sel-finance"
+          value={financePage}
+          onChange={(event) => {
+            setFinancePage(event.target.value);
+          }}
+        >
+          <option value="sales">Sales</option>
+          <option value="expenses">Expenses</option>
+          <option value="profits">Profits</option>
         </select>
       ) : (
         ""
