+++ src/layouts/admin.scss
@@ -83,6 +83,7 @@ body {
 .content-wrapper {
   margin-top: 68px;
   padding-top: 30px;
+  overflow-y: auto;
 }
 
 .sidebar {
@@ -143,7 +144,6 @@ body {
     height: 44px;
     width: 100%;
     display: block;
-    border: 1px solid #DFDFDF;
     border-radius: 6px;
     margin-bottom: 10px;
 
@@ -171,7 +171,7 @@ body {
   margin-bottom: 30px;
 }
 
-.common-button.active {
+.common-button:hover {
   background: linear-gradient(180deg, #8B5E83 20.45%, #82567A 100%, #82567A 100%);
   box-shadow: 0px 20px 24px rgba(70, 6, 81, 0.08);
   color: white;
