+++ src/pages/ERP/Project/Components/ProjectSchedule/calendar.css
@@ -1,7 +1,6 @@
 .react-calendar {
     display: inline-block;
-    width: 380px;
-    padding: 15px;
+    padding: 16px;
     border-radius: 5px;
     max-width: 100%;
     background: white;
@@ -87,7 +86,7 @@
 
   .calendar-select {
     height: 30px;
-    margin: 5px 7px;
+    margin-left: 15px;
     border: 1px solid #C8CDD6;
     border-radius: 3px;
     padding: 5px 10px;
